+++ Runtime/Scripts/GlobalTools/UI/ProgressBar.cs
@@ -7,14 +7,18 @@ using PolytopeSolutions.Toolset.GlobalTools.Generic;
 
 namespace PolytopeSolutions.Toolset.GlobalTools.UI {
     public class ProgressBar : MonoBehaviour {
-        [SerializeField] private Image progressBarImage;
+        [SerializeField] protected Image progressBarImage;
+        [SerializeField] protected float smoothDuration = 0.1f;
 
-        protected virtual float progress {
+        protected virtual float targetProgress {
             get { return 0f; }
         }
+        private float currentProgress;
+        private float currentVelocity;
 
-        void Update() { 
-            this.progressBarImage.fillAmount = this.progress;
+        private void Update() { 
+            this.currentProgress = Mathf.SmoothDamp(this.currentProgress, this.targetProgress, ref this.currentVelocity, this.smoothDuration);
+            this.progressBarImage.fillAmount = this.currentProgress;
         }
     }
 }